refactor: Move `NodeCmd` and `NodeArgs` into `cmd` module
This was a pretty clean move, and it helps highlight areas where
responsibility may be poorly assigned (e.g. duplicate imports).
