+++ src/lib.rs
@@ -7,24 +7,25 @@
 // specific language governing permissions and limitations relating to use of the SAFE Network
 // Software.
 
+mod cmd;
+
 use eyre::{eyre, Result, WrapErr};
 use std::fs;
 use std::{
     borrow::Cow,
     collections::HashSet,
     env,
-    ffi::{OsStr, OsString},
-    fmt,
     fs::File,
     io::{self, BufReader},
     net::SocketAddr,
     path::{Path, PathBuf},
-    process::{Command, Stdio},
     thread,
     time::Duration,
 };
 use structopt::StructOpt;
-use tracing::{debug, info, trace};
+use tracing::{debug, info};
+
+use cmd::NodeCmd;
 
 #[cfg(not(target_os = "windows"))]
 const SN_NODE_EXECUTABLE: &str = "sn_node";
@@ -36,7 +37,6 @@ const SN_NODE_EXECUTABLE: &str = "sn_node.exe";
 const GENESIS_CONN_INFO_FILEPATH: &str = ".safe/node/node_connection_info.config";
 
 const DEFAULT_RUST_LOG: &str = "safe_network=debug";
-const NODE_LIVENESS_TIMEOUT: Duration = Duration::from_secs(2);
 
 /// Tool to launch Safe nodes to form a local single-section network
 ///
@@ -247,7 +247,7 @@ fn launch(args: &Launch) -> Result<()> {
 
     debug!(
         "Common node args for launching the network: {:?}",
-        node_cmd.args
+        node_cmd.args()
     );
 
     let paths =
@@ -301,7 +301,10 @@ fn join(args: &Join) -> Result<()> {
         return Ok(());
     }
 
-    debug!("Node to be started with contact(s): {:?}", args.hard_coded_contacts);
+    debug!(
+        "Node to be started with contact(s): {:?}",
+        args.hard_coded_contacts
+    );
 
     debug!("Launching node...");
     node_cmd.run(&args.nodes_dir, &args.hard_coded_contacts)?;
@@ -313,172 +316,6 @@ fn join(args: &Join) -> Result<()> {
     debug!("(Note that log files are rotated hourly, and subsequent files will be named sn_node.log<NEW DATE TINE>.");
 
     Ok(())
-}
-
-#[derive(Clone)]
-struct NodeCmd<'a> {
-    path: Cow<'a, OsStr>,
-    envs: Vec<(Cow<'a, OsStr>, Cow<'a, OsStr>)>,
-    args: NodeArgs<'a>,
-}
-
-impl<'a> NodeCmd<'a> {
-    fn new<P, Pb>(path: P) -> Self
-    where
-        P: Into<Cow<'a, Pb>>,
-        Pb: AsRef<OsStr> + ToOwned + ?Sized + 'a,
-        Pb::Owned: Into<OsString>,
-    {
-        Self {
-            path: into_cow_os_str(path),
-            envs: Default::default(),
-            args: Default::default(),
-        }
-    }
-
-    fn path(&self) -> &Path {
-        Path::new(&self.path)
-    }
-
-    fn push_env<K, Kb, V, Vb>(&mut self, key: K, value: V)
-    where
-        K: Into<Cow<'a, Kb>>,
-        Kb: AsRef<OsStr> + ToOwned + ?Sized + 'a,
-        Kb::Owned: Into<OsString>,
-        V: Into<Cow<'a, Vb>>,
-        Vb: AsRef<OsStr> + ToOwned + ?Sized + 'a,
-        Vb::Owned: Into<OsString>,
-    {
-        self.envs
-            .push((into_cow_os_str(key), into_cow_os_str(value)));
-    }
-
-    fn push_arg<A, B>(&mut self, arg: A)
-    where
-        A: Into<Cow<'a, B>>,
-        B: AsRef<OsStr> + ToOwned + ?Sized + 'a,
-        B::Owned: Into<OsString>,
-    {
-        self.args.push(arg);
-    }
-
-    fn print_version(&self) -> Result<()> {
-        let version = Command::new(&self.path)
-            .args(&["-V"])
-            .output()
-            .map_or_else(
-                |error| Err(eyre!(error)),
-                |output| {
-                    if output.status.success() {
-                        Ok(output.stdout)
-                    } else {
-                        Err(eyre!(
-                            "Process exited with non-zero status (status: {}, stderr: {})",
-                            output.status,
-                            String::from_utf8_lossy(&output.stderr)
-                        ))
-                    }
-                },
-            )
-            .wrap_err_with(|| {
-                format!(
-                    "Failed to run '{}' with args '{:?}'",
-                    self.path().display(),
-                    &["-V"]
-                )
-            })?;
-
-        debug!(
-            "Using sn_node @ {} from {}",
-            String::from_utf8_lossy(&version).trim(),
-            self.path().display()
-        );
-
-        Ok(())
-    }
-
-    fn run(&self, node_dir: impl AsRef<Path>, contacts: &[SocketAddr]) -> Result<()> {
-        let path_str = self.path().display().to_string();
-        trace!("Running '{}' with args {:?} ...", path_str, self.args);
-
-        let mut extra_args = NodeArgs::default();
-        extra_args.push("--root-dir");
-        extra_args.push(node_dir.as_ref());
-        extra_args.push("--log-dir");
-        extra_args.push(node_dir.as_ref());
-
-        if !contacts.is_empty() {
-            extra_args.push("--hard-coded-contacts");
-            extra_args.push(
-                serde_json::to_string(
-                    &contacts
-                        .iter()
-                        .map(|contact| contact.to_string())
-                        .collect::<Vec<_>>(),
-                )
-                .wrap_err("Failed to generate genesis contacts list parameter")?,
-            );
-        }
-
-        Command::new(&path_str)
-            .args(&self.args)
-            .args(&extra_args)
-            .envs(self.envs.iter().map(
-                // this looks like a no-op but really converts `&(_, _)` into `(_, _)`
-                |(key, value)| (key, value),
-            ))
-            .stdout(Stdio::null())
-            .stderr(Stdio::inherit())
-            .spawn()
-            .map_err(|error| eyre!(error))
-            .and_then(|mut child| {
-                // Wait a couple of seconds to see if the node fails immediately, so we can fail fast
-                thread::sleep(NODE_LIVENESS_TIMEOUT);
-
-                if let Some(status) = child.try_wait()? {
-                    return Err(eyre!("Node exited early (status: {})", status));
-                }
-
-                Ok(())
-            })
-            .wrap_err_with(|| {
-                format!("Failed to start '{}' with args '{:?}'", path_str, self.args)
-            })?;
-
-        Ok(())
-    }
-}
-
-#[derive(Clone, Default)]
-struct NodeArgs<'a>(Vec<Cow<'a, OsStr>>);
-
-impl<'a> NodeArgs<'a> {
-    fn push<A, B>(&mut self, arg: A)
-    where
-        A: Into<Cow<'a, B>>,
-        B: AsRef<OsStr> + ToOwned + ?Sized + 'a,
-        B::Owned: Into<OsString>,
-    {
-        self.0.push(into_cow_os_str(arg));
-    }
-}
-
-impl<'a> IntoIterator for &'a NodeArgs<'a> {
-    type Item = &'a Cow<'a, OsStr>;
-
-    type IntoIter = std::slice::Iter<'a, Cow<'a, OsStr>>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.iter()
-    }
-}
-
-impl<'a> fmt::Debug for NodeArgs<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_list()
-            .entries(self.0.iter().map(|arg| -> &OsStr { arg.as_ref() }))
-            .finish()
-    }
 }
 
 fn read_genesis_conn_info() -> Result<Vec<SocketAddr>> {
@@ -506,16 +343,4 @@ fn read_genesis_conn_info() -> Result<Vec<SocketAddr>> {
     debug!("Genesis node contact info: {:?}", contacts);
 
     Ok(contacts)
-}
-
-fn into_cow_os_str<'a, V, Vb>(val: V) -> Cow<'a, OsStr>
-where
-    V: Into<Cow<'a, Vb>>,
-    Vb: AsRef<OsStr> + ToOwned + ?Sized + 'a,
-    Vb::Owned: Into<OsString>,
-{
-    match val.into() {
-        Cow::Borrowed(val) => val.as_ref().into(),
-        Cow::Owned(val) => val.into().into(),
-    }
 }
